requirements.txt
arrow==1.1.0
asgiref==3.3.4
binaryornot==0.4.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
cryptography==3.4.7
Django==3.2
django-tailwind==2.0.1
docker==5.0.0
future==0.18.2
idna==2.10
Jinja2==2.11.3
jinja2-time==0.2.0
josepy==1.8.0
MarkupSafe==1.1.1
mozilla-django-oidc==1.2.4
poyo==0.5.0
pycparser==2.20
pycryptodomex==3.10.1
pyjwkest==1.4.2
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-slugify==5.0.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.4
websocket-client==0.58.0
channels
graphql-ws
